Fix type of evnt_multi() mepbuff parameter
Header was WORD *mepbuff while implementation was BYTE *mepbuff.

+++ desk/gembind.c
@@ -343,7 +343,7 @@ WORD evnt_timer(UWORD locnt, UWORD hicnt)
 WORD evnt_multi(UWORD flags, UWORD bclk, UWORD bmsk, UWORD bst,
                 UWORD m1flags, UWORD m1x, UWORD m1y, UWORD m1w, UWORD m1h,
                 UWORD m2flags, UWORD m2x, UWORD m2y, UWORD m2w, UWORD m2h,
-                BYTE *mepbuff, UWORD tlc, UWORD thc, UWORD *pmx, UWORD *pmy,
+                WORD *mepbuff, UWORD tlc, UWORD thc, UWORD *pmx, UWORD *pmy,
                 UWORD *pmb, UWORD *pks, UWORD *pkr, UWORD *pbr )
 {
         MU_FLAGS = flags;
